Security flaw: CryptoKit was using random nonces.
Perhaps that was obvious, but the documentation states, incorrectly,
that creating a new nonce with an empty initializer is safe.

That's not safe: it's generating a random nonce, which has a non-
negligble chance of collisions, because the nonce is only 96 bits.
A nonce collision ruins the security of CTR.

I'm now manually setting the nonce and using them to count, as CTR
expects. That took more code than I wanted to write for low-level
cryptographic operations.

Boo on Apple, this really shouldn't be possible. Here we are in 2023
and we still have to contend with unsafe crypto APIs.

grrr argh